+++ Vine/Classes/Vine.swift
@@ -55,80 +55,3 @@ open class Vine<Root: AnyObject>: VineType {
         startFunction?(self)
     }
 }
-
-// MARK: - Window
-
-public class Window: UIWindow {
-    /// Reference to the Vine
-    var vine: Vine<Window>
-
-    public init(frame: CGRect, vine: Vine<Window>) {
-        self.vine = vine
-        super.init(frame: frame)
-        self.vine.root = self
-        self.vine.start()
-    }
-
-    public required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-// MARK: - NavigationController
-
-// The initializers for this are a little fucked up at the moment.
-public class NavigationController: UINavigationController {
-    public typealias VineType = Vine<NavigationController>
-    var vine: VineType?
-
-    public override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
-        super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
-    }
-
-    public init(vine: VineType, navigationBarClass: AnyClass? = nil, toolbarClass: AnyClass? = nil) {
-        super.init(navigationBarClass: navigationBarClass, toolbarClass: toolbarClass)
-        self.vine = vine
-        self.vine?.root = self
-        self.vine?.start()
-    }
-
-    public required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-// MARK: - SplitViewController
-
-public class SplitViewController: UISplitViewController {
-    public typealias VineType = Vine<SplitViewController>
-    var vine: VineType
-
-    public init(nibName nibNameOrNil: String? = nil, bundle nibBundleOrNil: Bundle? = nil, vine: VineType) {
-        self.vine = vine
-        super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
-        self.vine.root = self
-        self.vine.start()
-    }
-
-    public required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-// MARK: - TabBarController
-
-public class TabBarController: UITabBarController {
-    public typealias VineType = Vine<TabBarController>
-    var vine: VineType
-
-    public init(nibName nibNameOrNil: String? = nil, bundle nibBundleOrNil: Bundle? = nil, vine: VineType) {
-        self.vine = vine
-        super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
-        self.vine.root = self
-        self.vine.start()
-    }
-
-    public required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
